Turned on Nullable Reference Types because this will save a lot of bother later.  Added Collider generic classes.  Removed Collision static methods.  Removed unnecessary new-ings and initialisations  Made some stuff readonly, internal and private where it can be / should be  Introduced Settings.Tau to replace all instances of 2*Pi
