+++ src/box2dx/Box2D.NetStandard/Box2D.NetStandard.csproj
@@ -13,6 +13,7 @@
     <FileVersion>2.0.0.0</FileVersion>
     <Version>2.0-alpha</Version>
     <PackageVersion>2.0-alpha</PackageVersion>
+    <Nullable>enable</Nullable>
   </PropertyGroup>
 
   <PropertyGroup Condition="'$(Configuration)|$(Platform)'=='Debug|AnyCPU'">
