+++ src/box2dx/Box2D.NetStandard/Collision/ContactFeature.cs
@@ -1,0 +1,11 @@
+namespace Box2D.NetStandard.Collision {
+  /// The features that intersect to form the contact point
+  /// This must be 4 bytes or less.
+  struct ContactFeature
+  {
+    internal byte indexA;		///< Feature index on shapeA
+    internal byte indexB;		///< Feature index on shapeB
+    internal byte typeA;		///< The feature type on shapeA
+    internal byte typeB; ///< The feature type on shapeB
+  };
+}
