+++ src/box2dx/Box2D.NetStandard/Collision/ContactFeatureType.cs
@@ -1,0 +1,7 @@
+namespace Box2D.NetStandard.Collision {
+  enum ContactFeatureType:byte
+  {
+    Vertex = 0,
+    Face   = 1
+  };
+}
