+++ src/box2dx/Box2D.NetStandard/Common/Settings.cs
@@ -31,6 +31,7 @@ namespace Box2D.NetStandard.Common
 {
 	public class Settings
 	{
+		
 		public const float AABBMultiplier = 4.0f;
 		public const float AABBExtension = 0.1f;
 		public const int MaxTOIContacts = 32;
@@ -41,16 +42,18 @@ namespace Box2D.NetStandard.Common
 		public static float	FORCE_SCALE2(x){ return x<<7;}
 		public static float FORCE_INV_SCALE2(x)	{return x>>7;}
 #else
-		
-		public const float FLT_EPSILON = 1.192092896e-07F;//smallest such that 1.0f+FLT_EPSILON != 1.0f
+
+		public const float FLT_EPSILON = float.Epsilon; //1.192092896e-07F;//smallest such that 1.0f+FLT_EPSILON != 1.0f
 		public const float FLT_EPSILON_SQUARED = FLT_EPSILON * FLT_EPSILON;//smallest such that 1.0f+FLT_EPSILON != 1.0f
 		public static float FORCE_SCALE(float x) { return x; }
 		public static float FORCE_INV_SCALE(float x) { return x; }
 #endif
 
-		[Obsolete("Use MathF.Pi instead",true)]
-		public const float Pi = 3.14159265359f;
-
+		public const float Pib2 = 3.14159265359f; // Original code. Comes out at 3.1415927f
+		public const float Pi = MathF.PI; // Also 3.1415927f
+		public const float Pi2 = (float) System.Math.PI; // Also displayed as "3.1415927f"
+		public const float Tau = 2f * Pi;
+		
 		// Global tuning constants based on meters-kilograms-seconds (MKS) units.
 
 		// Collision
@@ -71,7 +74,7 @@ namespace Box2D.NetStandard.Common
 		/// A small angle used as a collision and constraint tolerance. Usually it is
 		/// chosen to be numerically significant, but visually insignificant.
 		/// </summary>
-		public const float AngularSlop = 2.0f / 180.0f * MathF.PI; // 2 degrees
+		public const float AngularSlop = 2.0f / 180.0f * Settings.Pi; // 2 degrees
 
 		/// <summary>
 		/// The radius of the polygon/edge shape skin. This should not be modified. Making
@@ -113,7 +116,7 @@ namespace Box2D.NetStandard.Common
 		/// The maximum angular position correction used when solving constraints.
 		/// This helps to prevent overshoot.
 		/// </summary>
-		public const float MaxAngularCorrection = 8.0f / 180.0f * MathF.PI; // 8 degrees
+		public const float MaxAngularCorrection = 8.0f / 180.0f * Settings.Pi; // 8 degrees
 
 		/// <summary>
 		/// The maximum linear velocity of a body. This limit is very large and is used
@@ -145,7 +148,7 @@ namespace Box2D.NetStandard.Common
 		/// The maximum angular velocity of a body. This limit is very large and is used
 		/// to prevent numerical problems. You shouldn't need to adjust this.
 		/// </summary>
-		public const float MaxRotation = (0.5f * MathF.PI);
+		public const float MaxRotation = (0.5f * Settings.Pi);
 		public const float MaxRotationSquared = (MaxRotation * MaxRotation);
 
 		/// <summary>
