+++ src/box2dx/Box2D.NetStandard/Dynamics/Contacts/CircleContact.cs
@@ -34,6 +34,8 @@ namespace Box2D.NetStandard.Dynamics.Contacts
 {
 	public class CircleContact : Contact
 	{
+		private static Collider<CircleShape,CircleShape> collider = new CircleAndCircleCollider();
+		
 		public CircleContact(Fixture fixtureA, Fixture fixtureB)
 			: base(fixtureA,0, fixtureB,0)
 		{
@@ -52,7 +54,9 @@ namespace Box2D.NetStandard.Dynamics.Contacts
 		}
 
 		internal override void Evaluate(out Manifold manifold, in Transform xfA, in Transform xfB) {
-			Collision.Collision.CollideCircles(out manifold,(CircleShape)m_fixtureA.Shape,xfA,(CircleShape)m_fixtureB.Shape,xfB);
+			
+			
+			collider.Collide(out manifold,(CircleShape)m_fixtureA.Shape,xfA,(CircleShape)m_fixtureB.Shape,xfB);
 		}
 	}
 }
