+++ src/box2dx/Box2D.NetStandard/Dynamics/Joints/Distance/DistanceJoint.cs
@@ -145,7 +145,7 @@ namespace Box2D.NetStandard.Dynamics.Joints.Distance {
         float C = length - _length;
 
         // Frequency
-        float omega = 2.0f * MathF.PI * _frequencyHz;
+        float omega = Settings.Tau * _frequencyHz;
 
         // Damping coefficient
         float d = 2.0f * _mass * _dampingRatio * omega;
