+++ tests/unit/Box2D.NetStandard.UnitTests/MathTests.cs
@@ -1,4 +1,3 @@
-using Box2D.NetStandard.Common;
 using Xunit;
 using Xunit.Abstractions;
